Move methods into HearingProcessor
